refactor: consolidate plan actions and remove unified directory
- Move IssueReviewButton from unified/ to registry/components/
- Flatten unified/types.ts into registry/types.ts
- Remove redundant getActionLabel helper
- Fix ISSUE_STATUS_REOPEN to only show for CANCELED issues (not DONE)
- Update ACTIONS.md documentation
