+++ frontend/src/components/Plan/components/HeaderSection/Actions/registry/components/index.ts
@@ -1,2 +1,3 @@
 export { default as ActionButton } from "./ActionButton.vue";
 export { default as ActionDropdown } from "./ActionDropdown.vue";
+export { default as IssueReviewButton } from "./IssueReviewButton.vue";
