--- frontend/src/components/Plan/components/HeaderSection/Actions/unified/index.ts
@@ -1,2 +1,0 @@
-export { default as IssueReviewButton } from "./IssueReviewButton.vue";
-export * from "./types";
